unifying connectionString and connectionLine in database and renaming files
